Separate TizenBuildInfo from tizen_builder.dart
